tests/integration/e2e/transfer_test.rs
use backend::mesh::service::MeshService;
use backend::mesh::offline::storage::OfflineStorage;
use backend::security::audit::AuditTrail;
use crate::common::helpers;
use std::time::Duration;

#[cfg(test)]
mod transfer_tests {
    use super::*;

    async fn setup_transfer_environment() -> TransferEnvironment {
        let mesh_service = MeshService::new();
        let offline_storage = OfflineStorage::new();
        let audit_trail = AuditTrail::new();
        
        TransferEnvironment {
            mesh_service,
            offline_storage,
            audit_trail,
        }
    }

    #[tokio::test]
    async fn test_secure_asset_transfer() {
        let env = setup_transfer_environment().await;
        
        // 1. Initialize transfer
        let transfer = create_secure_transfer();
        let init_result = env.mesh_service
            .initiate_transfer(transfer.clone())
            .await;
        assert!(init_result.is_ok(), "Transfer initiation should succeed");

        // 2. Verify authentication
        let auth_result = env.mesh_service
            .verify_transfer_auth(&transfer)
            .await;
        assert!(auth_result.is_ok(), "Transfer authentication should succeed");

        // 3. Execute transfer
        let execution_result = env.mesh_service
            .execute_transfer(transfer.clone())
            .await;
        assert!(execution_result.is_ok(), "Transfer execution should succeed");

        // 4. Verify audit trail
        let audit_entries = env.audit_trail
            .get_transfer_audit_trail(&transfer.id)
            .await;
        assert!(!audit_entries.is_empty(), "Audit trail should be recorded");
    }

    #[tokio::test]
    async fn test_emergency_transfer_protocol() {
        let env = setup_transfer_environment().await;
        
        // 1. Initiate emergency transfer
        let emergency_transfer = create_emergency_transfer();
        let result = env.mesh_service
            .execute_emergency_transfer(emergency_transfer.clone())
            .await;
        assert!(result.is_ok(), "Emergency transfer should succeed");

        // 2. Verify expedited processing
        let processing_time = result.unwrap().processing_time;
        assert!(processing_time < Duration::from_secs(1), 
                "Emergency transfer should be processed quickly");

        // 3. Verify proper authorization
        let auth_status = env.mesh_service
            .verify_emergency_auth(&emergency_transfer)
            .await;
        assert!(auth_status.is_valid(), "Emergency authorization should be valid");
    }

    #[tokio::test]
    async fn test_batch_transfer_operations() {
        let env = setup_transfer_environment().await;
        
        // 1. Create batch of transfers
        let transfers = create_batch_transfers(5);
        
        // 2. Execute batch transfer
        let batch_result = env.mesh_service
            .execute_batch_transfer(transfers.clone())
            .await;
        assert!(batch_result.is_ok(), "Batch transfer should succeed");

        // 3. Verify all transfers completed
        for transfer in transfers {
            let status = env.mesh_service
                .get_transfer_status(&transfer.id)
                .await;
            assert!(status.is_completed(), "All transfers should complete");
        }
    }

    // Helper structs and functions
    struct TransferEnvironment {
        mesh_service: MeshService,
        offline_storage: OfflineStorage,
        audit_trail: AuditTrail,
    }

    fn create_secure_transfer() -> AssetTransfer {
        AssetTransfer {
            id: format!("TRANSFER_{}", uuid::Uuid::new_v4()),
            asset_id: "ASSET_001".to_string(),
            source_unit: "UNIT_A".to_string(),
            destination_unit: "UNIT_B".to_string(),
            security_level: SecurityLevel::High,
            timestamp: chrono::Utc::now(),
        }
    }

    fn create_emergency_transfer() -> EmergencyTransfer {
        EmergencyTransfer {
            id: format!("EMERG_{}", uuid::Uuid::new_v4()),
            asset_id: "ASSET_001".to_string(),
            reason: EmergencyReason::UnitRelocation,
            authorization_code: "EMERG_AUTH_123".to_string(),
            timestamp: chrono::Utc::now(),
        }
    }

    fn create_batch_transfers(count: u32) -> Vec<AssetTransfer> {
        (0..count)
            .map(|i| AssetTransfer {
                id: format!("TRANSFER_{}", i),
                asset_id: format!("ASSET_{}", i),
                source_unit: "UNIT_A".to_string(),
                destination_unit: "UNIT_B".to_string(),
                security_level: SecurityLevel::Standard,
                timestamp: chrono::Utc::now(),
            })
            .collect()
    }
}
